Find gs executable across all platforms
The existing config_platform is broken on newer Apple Silicon-based
macOS, where Homebrew will install `gs` to /opt/homebrew/...

This introduces a platform-agnostic equivalent of Unix's `which` to find
an executable along the $PATH entries. It should also work with Windows
- but I don't have a Windows machine to verify that on. Where we can, we
rely on Ruby's built-in abstractions for file access and path building.

With this change the specs are all green on my M1 Max machine.